refactor(SnackbarContent): rename files to match component
