+++ packages/core/src/components/Snackbar/SnackbarContent/SnackbarContent.styles.ts
