+++ packages/core/src/components/Snackbar/SnackbarContent/index.ts
@@ -1,0 +1,1 @@
+export * from "./SnackbarContent";
