--- packages/core/src/components/Snackbar/SnackbarContentWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from "./SnackbarContentWrapper";
